Initial commit into the new repository.
With all the struggles with imports and Python infrastructure, the old repository was a complete mess, and included too many curse words. 😹 After much trial and error, mostly errors, I was able to get tbp building and running as a module. I apologize for not having the history for the project.