The docs now show the version number.
The change turned out to be simple, just adding `./docs/_data/version.yml` file with the version number in it. Additionally, added `./docs/_include/nav_footer_custom.html`, which is a simple Liquid tag to read the version file.
This change will necessitate coming up with a solution to keep the code version number in .`/src/tbp/__init__.py` and the new YAML file in sync.
Also updated the `makefile` to use `bundle exec` in front of the documentation commands. That was not clearly spelled out in the Jekyll docs.

Closes #4.

Adding lots of diagnostics

In trying to figure out why upload actions failed, added more diagnostics and explicitly set names for the coverage files. Also set up workflow to ignore any changes in the docs directory.

Fixing coverage combine

I renamed the output coverage files, but forgot to update the "Combine and Report" step in `cov-report-job`

Coverage.py really wants files to be named .coverage

Changed all file creation and usage to use .coverage*