The docs now show the version number.
The change turned out to be simple, just adding `./docs/_data/version.yml` file with the version number in it. Additionally, added `./docs/_include/nav_footer_custom.html`, which is a simple Liquid tag to read the version file.
This change will necessitate coming up with a solution to keep the code version number in .`/src/tbp/__init__.py` and the new YAML file in sync.
Also updated the `makefile` to use `bundle exec` in front of the documentation commands. That was not clearly spelled out in the Jekyll docs.

Closes #4.